% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.R
\name{fv}
\alias{fv}
\title{Returns the future value of an investment (fv)}
\usage{
fv(rate = 0, nper = 1, pv = 0, pmt = 0, pmtUltimo = TRUE)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal}

\item{nper}{The total number of payment periods. Default is one period}

\item{pv}{The present value of single investment made today. Default is assumed to be zero. Must be entered as a negative number}

\item{pmt}{The payment made each period (annuity). Must be entered as a negative number.}

\item{pmtUltimo}{When payments are due. TRUE = end of period, FALSE = beginning of period. Default is TRUE.}
}
\description{
Returns the future value of an investment (fv)
}
\examples{
fv(rate=0.01,nper=10,pv=1000,pmt=10,pmtUltimo=TRUE)
}
\seealso{
\code{\link{fv.single}}

\code{\link{fv.annuity}}
}
