% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fv.single.R
\name{fv.single}
\alias{fv.single}
\title{Returns the future value of an single payment (fv)}
\usage{
fv.single(rate = 0, inflation = 0, nper = 1, pv = 0)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal or ts}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal or ts}

\item{nper}{The total number of payment periods. Default is one period. If rate and inflation are entered as ts nper is ignored.}

\item{pv}{The present value of single payment made today. Default is assumed to be zero. Must be entered as a negative number}
}
\description{
Returns the future value of an single payment (fv)
}
\examples{
fv(rate=0.01,nper=10,pv=-1000)
fv.single(rate=0.01,nper=10,pv=-1000)
}
\seealso{
\code{\link{fv}}
}
