% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gompertzMakehamMortality.R
\name{gompertzMakehamMortality}
\alias{gompertzMakehamMortality}
\title{Calculates a variety of morality parameters given by Gompertz-Makeham Law of Mortality}
\usage{
gompertzMakehamMortality(
  x,
  lambda,
  m,
  b,
  t = 0,
  r = 0,
  sigma = 0,
  inflation = 0,
  B = 0,
  K = 0
)
}
\arguments{
\item{x}{Annuitant’s age x in years.}

\item{lambda}{Gompertz-Makeham accidental death rate, often set to 0.}

\item{m}{Gompertz-Makeham  modal natural death rate.}

\item{b}{Gompertz-Makeham dispersion of natural death rate.}

\item{t}{number of additional years}

\item{r}{Interest rate r (annual compounding)}

\item{sigma}{Standard deviation of annualized real log returns}

\item{inflation}{Inflation rate i (annual compounding)}

\item{B}{Constant continuous real spending at rate}

\item{K}{Wealth at retirement}
}
\description{
Calculates a variety of morality parameters given by Gompertz-Makeham Law of Mortality
}
\examples{
gompertzMakehamMortality(x=65,lambda=0,m=82.3,b=11.4,t=10,r=0.01,sigma=0,inflation=0,B=100,K=1000)
}
