% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infladj.R
\name{infladj}
\alias{infladj}
\title{Returns the adjusted future value of an investment (fv) for inflation}
\usage{
infladj(fv, inflation = 0, nper)
}
\arguments{
\item{fv}{The future value of an single payment (fv). Must be entered as a negative number}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal}

\item{nper}{The total number of payment periods. Default is one period}
}
\description{
Returns the adjusted future value of an investment (fv) for inflation
}
\examples{
infladj(fv=-1000,inflation=0.02, nper= 25)
}
