% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.scalar.R
\name{is.scalar}
\alias{is.scalar}
\title{Tests if an input is a scalar. Returns a boolean.}
\usage{
is.scalar(x)
}
\arguments{
\item{x}{Object to be tested}
}
\description{
Tests if an input is a scalar. Returns a boolean.
}
\examples{
is.scalar(0.04)
is.scalar(4)
is.scalar(c(1:3))
}
