% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximumSpendingForMinimumRuinTimeV2.R
\name{maximumSpendingForMinimumRuinTimeV2}
\alias{maximumSpendingForMinimumRuinTimeV2}
\title{Calculates scenarios of future value of annuity payments (fv) with stochastic returns}
\usage{
maximumSpendingForMinimumRuinTimeV2(wealth = 14000,
  minumumRuinTime = 16, mu = 0.03, sigma = 0.08, nScenarios = 200,
  prob = 0.9, seed = NULL)
}
\arguments{
\item{wealth}{The wealth at retirement. Must be entered as a positive number}

\item{minumumRuinTime}{Minimum time to ruin.  Must be entered as a positive integer}

\item{mu}{The expected interest real return per period. Default is zero. Must be entered as decimal}

\item{sigma}{Volatility of expected interest real return per period. Default is zero. Must be entered as decimal}

\item{nScenarios}{The total number of scenarios to be made. Default is one scenario}

\item{prob}{Probability to exceed minimum time to ruin. Must be entered as decimal.}

\item{seed}{Integer vector, containing the random number generator (RNG) state for random number generation in R}
}
\description{
Calculates scenarios of future value of annuity payments (fv) with stochastic returns
}
\examples{
maximumSpendingForMinimumRuinTimeV2(wealth=14000,minumumRuinTime = 16,mu=0.03,sigma=0.08,nScenarios=200, prob = 0.9, seed =NULL)

}
