% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payments.convert.R
\name{payment.convert}
\alias{payment.convert}
\title{Returns the future value of annuity payments (fv)}
\usage{
payment.convert(inflation, nper, pmt, pmtNominal = FALSE)
}
\description{
Returns the future value of annuity payments (fv)
}
\examples{
fv.annuity(rate=0.01,nper=10,pmt=-10,pmtUltimo=TRUE)
}
\seealso{
\code{\link{fv}}
}
