% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.growth.R
\name{pmt.growth}
\alias{pmt.growth}
\title{Returns timeseris of payments with growth}
\usage{
pmt.growth(nper = 1, pmt = 0, pmtGrowth = 0, pmtinfladj = FALSE, inflation = 0)
}
\arguments{
\item{nper}{The total number of payment periods. Default is one period.}

\item{pmt}{The payment (spending) made each period (annuity) in the future. Must be entered as a negative number.}

\item{pmtGrowth}{The payment (spending) Growth from one period to another.}

\item{pmtinfladj}{Should the payments be inflation adjusted? E.g. are the annuity pmt constant or real annuities. Only avaliable for pmt given as scalar. Default value = FALSE.}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal or ts}
}
\description{
Returns timeseris of payments with growth
}
\examples{
pmt.growth(nper=10,pmt=10,pmtGrowth=0.1)
}
