% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmt.R
\name{pmt}
\alias{pmt}
\title{Returns the payment (eg. on loan) based on constant payments and a constant interest rate}
\usage{
pmt(rate = 0.04, inflation = 0.02, nper = 35, fv = -1000)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal or ts}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal or ts}

\item{nper}{The total number of payment periods. Default is one period. If rate and inflation are entered as ts nper is ignored.}

\item{fv}{The future value of single payment (spending) made in the future. Default is assumed to be zero. Must be entered as a negative number}
}
\description{
Returns the payment (eg. on loan) based on constant payments and a constant interest rate
}
\examples{
pmt(rate=0.05,nper=35,fv=-1000)
pmt(rate=0.01,inflation=0,nper=10,fv=-1000)
}
\seealso{
\code{\link{pv}}
}
