% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.annuity.growth.R
\name{pv.annuity.growth}
\alias{pv.annuity.growth}
\title{Returns the present value of first payment of an annuit with growth based og present net worth}
\usage{
pv.annuity.growth(
  rate = 0,
  inflation = 0,
  nper = 1,
  netWorth = 1000,
  annuityGrowth = 0
)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal}

\item{nper}{The total number of payment periods. Default is one period}

\item{netWorth}{The payment (spending) made each period (annuity) in the future. Must be entered as a negative number.}

\item{annuityGrowth}{The annuity growth from one period to another.}
}
\description{
Returns the present value of first payment of an annuit with growth based og present net worth
}
\examples{
pv.annuity.growth(rate=0.01,inflation=0, nper=10,netWorth=1000, annuityGrowth=0.02)
}
