% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.values.R
\name{quantile.values}
\alias{quantile.values}
\title{Returns the specified quantiles of a numeric vector}
\usage{
\method{quantile}{values}(values, quantiles = c(0, 0.25, 0.5, 0.75, 1))
}
\arguments{
\item{values}{Numerical vector}

\item{quantiles}{Quantiles to be returned. Should be a numeric vector of probabilities with values in [0,1]}
}
\description{
Returns the specified quantiles of a numeric vector
}
\examples{
quantile.values(values=c(1:100), quantiles=c(0,0.25,0.5,0.75,1))
}
