% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.convert.R
\name{rate.convert}
\alias{rate.convert}
\title{Converts an rate with a given frequency to aneffective rate for another frequence. Base=1 is yearly rate}
\usage{
rate.convert(rate = 0, frequency.input = 1, frequency.output = 12)
}
\arguments{
\item{rate}{interest rate.}

\item{frequency.input}{Frequency of input rate. 1=annual, 12=monthly.}

\item{frequency.output}{Frequency of output rate. 1=annual, 12=monthly.}
}
\description{
Converts an rate with a given frequency to aneffective rate for another frequence. Base=1 is yearly rate
}
\examples{
rate.convert(rate=0.00327374,frequency.input=12,frequency.output=1) #Monthly to annual Rate
rate.convert(rate=0.04,frequency.input=1,frequency.output=12) #Annual to monthly Rate
}
