% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate.real.R
\name{rate.real}
\alias{rate.real}
\title{Converts an rate with a given frequency to aneffective rate for another frequence. Base=1 is yearly rate}
\usage{
rate.real(nominalRate, inflation)
}
\arguments{
\item{nominalRate}{nominal rate.}

\item{inflation}{inflation.}
}
\description{
Converts an rate with a given frequency to aneffective rate for another frequence. Base=1 is yearly rate
}
\examples{
rate.real(rate=0.00327374,inflation=0.02)
}
