% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requiredSavingsForMinimumAnnuity.R
\name{requiredSavingsForMinimumAnnuity}
\alias{requiredSavingsForMinimumAnnuity}
\title{Finds the required savings for minimum annuity}
\usage{
requiredSavingsForMinimumAnnuity(nper = 1, mu = 0, sigma = 0,
  convRate = 1, nScenarios = 1, minPayouy = 1000, prob = 0.95,
  seed = NULL, print = FALSE, resturnScenarios = FALSE)
}
\arguments{
\item{nper}{The total number of payment periods. Default is one period}

\item{mu}{The expected interest real return per period. Default is zero. Must be entered as decimal}

\item{sigma}{Volatility of expected interest real return per period. Default is zero. Must be entered as decimal}

\item{convRate}{The conversion rate. Default is one. Must be entered as decimal}

\item{nScenarios}{The total number of scenarios to be made. Default is one scenario}

\item{minPayouy}{The minimum desired yearly pension payout/target.}

\item{prob}{Probability to reach minimum desired yearly pension payout. Must be entered as decimal}

\item{seed}{Integer vector, containing the random number generator (RNG) state for random number generation in R}

\item{print}{Should the scenarios be displayed in plot}

\item{resturnScenarios}{Should the scenarios be returned in a matrix}
}
\description{
Finds the required savings for minimum annuity
}
\examples{
requiredSavingsForMinimumAnnuity(nper=1,mu=0,sigma=0,convRate=1,nScenarios=1,minPayouy = 1000, prob = 0.95, seed =NULL,print=FALSE,resturnScenarios=FALSE)
}
