% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarioGenerator.R
\name{scenarioGenerator}
\alias{scenarioGenerator}
\title{Makes scenarios with rnorm.}
\usage{
scenarioGenerator(nper = 25, nScenarios = 1000, mean = 0, sd = 1,
  seed = NULL)
}
\arguments{
\item{nper}{The total number of payment periods. Default is one period}

\item{nScenarios}{The total number of scenarios to be made. Default is one scenario}

\item{mean}{The mean of random normal distribution to sample from.}

\item{sd}{The standard deviation of random normal distribution to sample from.}

\item{seed}{Integer vector, containing the random number generator (RNG) state for random number generation in R}
}
\description{
Makes scenarios with rnorm.
}
\examples{
scenarioGenerator(nper=25,nScenarios=1000,mean=0,sd=1,seed=NULL)

}
