% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeToRuin.scenario.R
\name{timeToRuin.scenario}
\alias{timeToRuin.scenario}
\title{Calculates scenarios of future value of annuity payments (fv) with stochastic returns}
\usage{
timeToRuin.scenario(spending = 100, nper = 10, mu = 0, sigma = 0,
  wealth = 1000, nScenarios = 1, returnScenarios = FALSE,
  quantiles = c(0, 0.25, 0.5, 0.75, 1), seed = NULL)
}
\arguments{
\item{spending}{The annual spending. Must be given as negative number.}

\item{nper}{The planing horizon.}

\item{mu}{The expected interest real return per period. Default is zero. Must be entered as decimal}

\item{sigma}{Volatility of expected interest real return per period. Default is zero. Must be entered as decimal}

\item{wealth}{The wealth at retirement. Must be entered as a positive number}

\item{nScenarios}{The total number of scenarios to be made. Default is one scenario}

\item{returnScenarios}{Should the scenarios be returned}

\item{quantiles}{Quantile scenarios to be returned. Should be a numeric vector of probabilities with values in [0,1]}

\item{seed}{Integer vector, containing the random number generator (RNG) state for random number generation in R}
}
\description{
Calculates scenarios of future value of annuity payments (fv) with stochastic returns
}
\examples{
timeToRuin.scenario(spending=100,nper=10,mu=0.01,sigma=0.01,wealth=1000,nScenarios=1000, returnScenarios = FALSE,quantiles=c(0,0.25,0.5,0.75,1), seed =NULL)
}
