% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pv.single.R
\name{pv.single}
\alias{pv.single}
\title{Returns the future value of an single payment (fv)}
\usage{
pv.single(rate = 0, inflation = 0, nper = 1, fv = 0)
}
\arguments{
\item{rate}{The interest rate per period. Default is zero. Must be entered as decimal or ts}

\item{inflation}{The inflation rate per period. Default is zero. Must be entered as decimal or ts}

\item{nper}{The total number of payment periods. Default is one period. If rate and inflation are entered as ts nper is ignored.}

\item{fv}{The present value of single payment made today. Default is assumed to be zero. Must be entered as a negative number}
}
\description{
Returns the future value of an single payment (fv)
}
\examples{
pv.single(rate=0.05,inflation=0.03,nper=35,fv=-1000)
pv.single(rate=0.01,inflation=0,nper=10,fv=-1000)
}
\seealso{
\code{\link{fv}}
}
